+++ src/krm3/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'K-Tech team'
-__version__ = '2.5.2'
+__version__ = '2.6.0'
 __name__ = 'krm3'